requirements/prod.txt
aiodns==3.1.1 ; python_version >= "3.8" and python_version < "4.0"
aiohttp==3.9.1 ; python_version >= "3.8" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
astunparse==1.6.3 ; python_version >= "3.8" and python_version < "4.0"
async-lru==2.0.4 ; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.8" and python_version < "3.12.0"
asyncpg-trek==0.3.1 ; python_version >= "3.8" and python_version < "4"
asyncpg==0.29.0 ; python_version >= "3.8" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.8" and python_version < "4.0"
better-profanity==0.6.1 ; python_version >= "3.8" and python_version < "4.dev0"
braceexpand==0.1.7 ; python_version >= "3.8" and python_version < "4.0"
brotli==1.1.0 ; python_version >= "3.8" and python_version < "4.0"
cchardet==2.1.7 ; python_version < "3.10" and python_version >= "3.8"
cffi==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
cysystemd==1.5.4 ; python_version >= "3.8" and python_version < "4"
cython==3.0.6 ; python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32"
discord-ext-ipcx==0.2.1 ; python_version >= "3.8" and python_version < "4.0"
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus.git@8686b5d1bbc1d3c862292eb436ab630d6e9c9b53 ; python_version >= "3.8" and python_version < "4.0"
discord-py==2.3.2 ; python_version >= "3.8" and python_version < "4.0"
discord-py[speed]==2.3.2 ; python_version >= "3.8" and python_version < "4.0"
frozenlist==1.4.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.6 ; python_version >= "3.8" and python_version < "4.0"
import-expression==1.1.4 ; python_version >= "3.8" and python_version < "4.0"
importlib-metadata==6.8.0 ; python_version >= "3.8" and python_version < "3.10"
jishaku==2.5.2 ; python_version >= "3.8" and python_version < "4.0"
langcodes[data]==3.3.0 ; python_version >= "3.8" and python_version < "4.0"
language-data==1.1 ; python_version >= "3.8" and python_version < "4.0"
marisa-trie==0.7.8 ; python_version >= "3.8" and python_version < "4.0"
msgspec==0.18.4 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.8" and python_version < "4.0"
orjson==3.9.10 ; python_version >= "3.8" and python_version < "4.0"
prometheus-async==22.2.0 ; python_version >= "3.8" and python_version < "4.0"
prometheus-client==0.19.0 ; python_version >= "3.8" and python_version < "4.0"
psutil==5.9.6 ; python_version >= "3.8" and python_version < "4.0"
pycares==4.4.0 ; python_version >= "3.8" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.8" and python_version < "4.0"
pygit2==1.13.3 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
setuptools==69.0.2 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.8" and python_version < "3.11"
uvloop==0.19.0 ; python_version >= "3.8" and python_version < "4.0" and sys_platform != "win32"
wheel==0.42.0 ; python_version >= "3.8" and python_version < "4.0"
winloop==0.0.9 ; python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32"
wrapt==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
yarl==1.9.3 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.8" and python_version < "3.10"
